Service that fetches the balance from the user
Huge refactory of the bootstrap and the dependency injector methods, so language preference for the user is stored through a strategy pattern.

Also added the field defaultLocale to the languages.

Created methods in the CurrencyService to extract the amount in decimal number, and also in text mode depending on the current locale.