Big big HUGE refactoring.
Created CurrencyLayer api.

Split CurrencyService in 2: CurrencyService dedicated to fetch data related to the users and CurrencyConversionService dedicated to conversion and formatting

Fixed a lot of issues of unused code and errors. @rmrbest I'll try to comment the code changes so you can see. It will be by no means a critique, but a reminder of why we should be very strict with code quality.

Tests are passing now, but since Controllers are not covered, maybe some things will fail. @rmrbest and @Littlemad, please, tell me if something that was working before (related to currency mainly) has stopped working, so I can fix it.