Separated LotteriesDataService in two parts
LotteriesDataService holds the methods connecting to the api and writing to our database, while LotteryService holds the methods querying our database.

When there's a failure because there is missing data on the database that should be there, LotteryService tries to call LotteriesDataService to update the data. (Now only for nextJackpot).

@rmrbest there are some TODOs for you